+++ docs/src/App.tsx
@@ -63,6 +63,7 @@ const SnackbarDemoPage = page(() => import('./pages/components/widgets/SnackbarD
 const TreeDemoPage = page(() => import('./pages/components/widgets/TreeDemo'));
 const ChipDemoPage = page(() => import('./pages/components/widgets/ChipDemo'));
 const DropzoneDemoPage = page(() => import('./pages/components/widgets/DropzoneDemo'));
+const WebGLCanvasDemoPage = page(() => import('./pages/components/widgets/WebGLCanvasDemo'));
 const DateSelectorDemoPage = page(() => import('./pages/components/field/DateSelectorDemo'));
 const CodeBlockDemoPage = page(() => import('./pages/components/widgets/CodeBlockDemo'));
 const MarkdownDemoPage = page(() => import('./pages/components/widgets/MarkdownDemo'));
@@ -347,6 +348,10 @@ export function App() {
         <Route
           path='/markdown-demo'
           element={<MarkdownDemoPage />}
+        />
+        <Route
+          path='/webglcanvas-demo'
+          element={<WebGLCanvasDemoPage />}
         />
         <Route
           path='/prop-patterns'